# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Report System API",
        "description": "API documentation for the report system backend. Most school-scoped endpoints require an owner Sanctum token with the `permit:owner` ability.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/schools/{school}/aggregates": {
            "get": {
                "tags": [
                    "Aggregates"
                ],
                "summary": "List grade aggregates",
                "operationId": "ceb38738a6649bf4850f39203afb71ed",
                "parameters": [
                    {
                        "name": "school",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aggregates retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Aggregates retrieved successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Aggregate"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Aggregates"
                ],
                "summary": "Create a grade aggregate",
                "operationId": "1fc2482bfdf45dd3463137c93b29ef0a",
                "parameters": [
                    {
                        "name": "school",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "minScore",
                                    "maxScore",
                                    "grade",
                                    "remarks"
                                ],
                                "properties": {
                                    "minScore": {
                                        "type": "integer",
                                        "example": 80,
                                        "maximum": 100,
                                        "minimum": 0
                                    },
                                    "maxScore": {
                                        "type": "integer",
                                        "example": 100,
                                        "maximum": 100,
                                        "minimum": 0
                                    },
                                    "grade": {
                                        "type": "string",
                                        "example": "A",
                                        "maxLength": 10
                                    },
                                    "remarks": {
                                        "type": "string",
                                        "example": "Excellent",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Aggregate created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Aggregate created successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Aggregate"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/schools/{school}/aggregates/{aggregate}": {
            "put": {
                "tags": [
                    "Aggregates"
                ],
                "summary": "Update a grade aggregate",
                "operationId": "1c904663064f35ab2115eb7352411e29",
                "parameters": [
                    {
                        "name": "school",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "aggregate",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "minScore": {
                                        "type": "integer",
                                        "example": 70,
                                        "maximum": 100,
                                        "minimum": 0
                                    },
                                    "maxScore": {
                                        "type": "integer",
                                        "example": 79,
                                        "maximum": 100,
                                        "minimum": 0
                                    },
                                    "grade": {
                                        "type": "string",
                                        "example": "B",
                                        "maxLength": 10
                                    },
                                    "remarks": {
                                        "type": "string",
                                        "example": "Very Good",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Aggregate updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Aggregate updated successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Aggregate"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiError": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Something went wrong."
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "Aggregate": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "schoolId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "minScore": {
                        "type": "integer",
                        "example": 80
                    },
                    "maxScore": {
                        "type": "integer",
                        "example": 100
                    },
                    "grade": {
                        "type": "string",
                        "example": "A"
                    },
                    "remarks": {
                        "type": "string",
                        "example": "Excellent"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctum personal access token. Send as: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}